+++ locallib.php
@@ -1084,6 +1084,41 @@ function theme_boost_union_get_fontawesome_checks_templatecontext() {
 }
 
 /**
+ * Helper function to compose the title of an external admin page.
+ * This is adopted from /admin/settings.php and done to make sure that the external admin pages look as similar as possible
+ * to the standard admin pages.
+ *
+ * @param string $pagename The page's name.
+ *
+ * @return string
+ */
+function theme_boost_union_get_externaladminpage_title($pagename) {
+    global $SITE;
+
+    $title = $SITE->shortname.': ';
+    $title .= get_string('administration', 'core').': ';
+    $title .= get_string('appearance', 'core').': ';
+    $title .= get_string('themes', 'core').': ';
+    $title .= get_string('pluginname', 'theme_boost_union').': ';
+    $title .= $pagename;
+
+    return $title;
+}
+
+/**
+ * Helper function to compose the heading of an external admin page.
+ * This is adopted from /admin/settings.php and done to make sure that the external admin pages look as similar as possible
+ * to the standard admin pages.
+ *
+ * @return string
+ */
+function theme_boost_union_get_externaladminpage_heading() {
+    global $SITE;
+
+    return $SITE->fullname;
+}
+
+/**
  * Helper function which adds the CSS files from the fontawesome file area to the Moodle page.
  * This function uses the fontawesome cache definition, i.e. it does not load the files from the filearea directly.
  * It's meant to be called by theme_boost_union_before_standard_html_head() only.
@@ -1122,5 +1157,31 @@ function theme_boost_union_add_fontawesome_to_page() {
             // Add the CSS file to the page.
             $PAGE->requires->css($facssurl);
         }
+    }
+}
+
+/**
+ * Helper function which adds the CSS file from the flavour to the Moodle page.
+ * It's meant to be called by theme_boost_union_before_standard_html_head() only.
+ * *
+ * @throws coding_exception
+ * @throws dml_exception
+ * @throws moodle_exception
+ */
+function theme_boost_union_add_flavourcss_to_page() {
+    global $CFG, $PAGE;
+
+    // Require flavours library.
+    require_once($CFG->dirroot . '/theme/boost_union/flavours/flavourslib.php');
+
+    // If any flavour applies to this page.
+    $flavour = theme_boost_union_get_flavour_which_applies();
+    if ($flavour != null) {
+        // Build the flavour CSS file URL.
+        $flavourcssurl = new moodle_url('/theme/boost_union/flavours/styles.php',
+                array('id' => $flavour->id, 'rev' => theme_get_revision()));
+
+        // Add the CSS file to the page.
+        $PAGE->requires->css($flavourcssurl);
     }
 }